+++ packages/orq-rc/RELEASES.md
@@ -10899,3 +10899,13 @@ Based on:
 - [python v4.15.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.43 - packages/orq-rc
+
+## 2026-09-14 10:13:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.44 - packages/orq-rc